feat: remove old mock repository implementations and add new mocks for user, task, and user role repositories
